Move MitmapImageModel
